requirements.txt
streamlit>=1.40
gspread>=6.1
google-auth>=2.35
google-api-python-client>=2.150
pandas>=2.2
